server/requirements.txt
google-api-python-client
pandas
nltk
scikit-learn
flask